backend/requirements.txt
fastapi
google-api-python-client
google-auth-oauthlib
pandas
plotly
pydantic-settings
tqdm
uvicorn
